requirements.txt
numpy==2.0.2
keras==3.6.0
scikit-learn==1.5.2
tensorflow==2.18.0
pillow==11.0.0
streamlit
